BrenierTwoFluid Library Functions

BrenierTwoFluid.CostCollection — Type

CostCollection

Collection of cost matrices involving two point clouds `X` and `Y`. 

Fields:
- `C_xy::CT`: N × M `AbstractMatrix`.
- `C_yx::CT`: M × N `AbstractMatrix`.
- `C_xx::CT`: N × N `AbstractMatrix`.
- `C_yy::CT`: M × M `AbstractMatrix`.

BrenierTwoFluid.LazyCost — Type

LazyCost

AbstractMatrix type that represents a cost matrix between two point clouds `x` and `y`. The cost matrix is not computed until it is indexed.

Fields:
- `x::AT`: N × d `AbstractArray` of points.
- `y::AT`: M × d `AbstractArray` of points.
- `c::FT`: `Callable` scalar cost function `x,y -> c(x,y)`. The default is 1/2 times the squared Euclidean distance.

BrenierTwoFluid.LazyGibbsKernel — Type

LazyGibbsKernel

AbstractMatrix type that represents a Gibbs kernel between two point clouds `x` and `y`. The Gibbs kernel is not computed until it is indexed.

Fields:
- `C::LazyCost`: N × M `LazyCost` matrix.
- `ε::Base.RefValue`: reference of the Gibbs kernel scale.

BrenierTwoFluid.SinkhornBarycenter — Type

SinkhornBarycenter

Represents a Sinkhorn barycenter problem.

Fields:
- `ω::Vector{T}`: weights of the input distributions.
- `Ss::Vector{SinkhornDivergence}`: Sinkhorn divergences between the input distributions and the barycenter.
- `∇c`: gradient of the cost function.
- `max_it::Int`: maximum number of iterations. This can be different from the maximum number of iterations of the Sinkhorn divergences, typically it is much smaller.
- `tol::T`: tolerance for the stopping criterion. This can be different from the tolerance of the Sinkhorn divergences.
- `δX::AT`: gradient of the barycenter problem with respect to the positions of the particles representing the barycenter.

Type parameters:
- `LOG, SAFE, SYM, ACC`: As in `SinkhornDivergence` and in fact identical to those values of the contained `SinkhornDivergence` objects.

BrenierTwoFluid.SinkhornDivergence — Type

SinkhornDivergence

Sinkhorn divergence between two probability distributions, represented as `SinkhornVariable`s.

Fields:
- `V1::SinkhornVariable`: first probability distribution.
- `V2::SinkhornVariable`: second probability distribution.
- `CC::CostCollection`: collection of cost matrices.
- `params::SinkhornParameters`: Sinkhorn algorithm parameters.

Type parameters:
- `LOG`: whether or not the Sinkhorn algorithm is performed in the log domain.
- `SAFE`, `SYM`, `ACC`, `DEB`: as in `SinkhornParameters`.
Default: log = anything, safe = true, sym = false, acc = anything, deb = true, lowrank = false

BrenierTwoFluid.SinkhornParameters — Type

SinkhornParameters

Fields:
- `Δ::T`: maximum value we expect the cost function to take
- `q::T`: scaling parameter. If ε-scaling is employed, the scale s is reduced by a factor q at each iteration.
- `η::T`: acceleration parameter. At every iteration, f ← (1-η)f₋ + ηf₊.
- `crit_it::Int`: after `crit_it` iterations, η is inferred from the change in of the marginal error and updated.
- `p_η::Int`: the change in marginal from `crit_it - p_η` to `crit_it` is used to infer η.
- `s::T`: current scale. When no ε-scaling is employed, this is constant and equal to ε.
- `ε::T`: minimum scale.
- `p::T`: the proportionality of ε to Δ (e.g. p = 2 for the quadratic cost)
- `tol::T`: tolerance for the stopping criterion, which is either the marginal error or the change in the dual potentials.
- `max_it::Int`: maximum number of iterations.
- `tol_it::Int`: the stopping criterion is checked every `tol_it` iterations.

Type parameters:
- `SAFE`: whether or not the stopping criterion is checked rather than just doing a certain amount of iterations. This is set to `false` only if `tol == Inf`.
- `SYM`: whether or not the Sinkhorn algorithm is symmetrized, i.e. f ← 0.5f₋ + 0.5f₊.
- `ACC`: whether or not the Sinkhorn algorithm is accelerated, i.e. f ← (1-η)f₋ + ηf₊.
- `DEB`: whether or not the Sinkhorn algorithm is debiased, i.e. computing W₂(μ,ν) - (W₂(μ,μ) + W₂(ν,ν))/2.

BrenierTwoFluid.SinkhornVariable — Type

SinkhornVariable

Probability distribution, represented as a collection of weighted particles, and its associated transport potentials.

Fields:
- `X::AT`: particle positions. `AT` is an N×d AbstractArray with elements of type `T`.
- `α::VT`: particle weights. `VT` is an N×1 AbstractVector with elements of type `T`.
- `log_α::VT`: logarithm of the particle weights.
- `f::VT`: transport potential, i.e. the dual variable associated with the probability distribution ∑ᵢαᵢδ(Xᵢ).
- `f₋::VT`: cache-like array to store the previous value of `f`.
- `h::VT`: transport potential for the de-biasing step.
- `h₋::VT`: cache-like array to store the previous value of `h`.
- `∇fh::AT`: gradient of `f+h` evaluated at `X`.

BrenierTwoFluid.compute! — Method

compute!

Solve the OT problem by performing Sinkhorn interations. If `S` is safe, the stopping criterion is checked rather than just doing a certain amount of iterations.

Arguments:
- `S::SinkhornDivergence`

Returns:
- `value(S)`: the value of the Sinkhorn divergence.

BrenierTwoFluid.marginal_error — Method

marginal_errors

Compute the second marginal error of the transport plan associated with `S`, ``\sum_j \pi_{ij} - α_i``, in the maximum norm.

Arguments:
- `S::SinkhornDivergence`

Returns:
- `π2_err`

BrenierTwoFluid.marginal_errors — Method

marginal_errors

Compute the marginal errors of the transport plan associated with `S`,``\sum_i \pi_{ij} - β_j`` and ``\sum_j \pi_{ij} - α_i``, in the maximum norm.

Arguments:
- `S::SinkhornDivergence`

Returns:
- `(π1_err, π2_err)`

BrenierTwoFluid.sinkhorn_step! — Method

sinkhorn_step!

Perform one iteration of the Sinkhorn algorithm, i.e. an update on both dual potentials as well as the two de-biasing potentials. This method dispatches to the correct implementation (log, safe, symmetric, accelerated) based on the type of `S`.

Arguments:
- `S::SinkhornDivergence`

BrenierTwoFluid.softmin — Method

softmin

Compute the softmin of `C[:,j] - f`, i.e. ``\log \sum_i α_i exp((f_i - C_{ij})/ε)`` in a numerically stable way (https://www.nowozin.net/sebastian/blog/streaming-log-sum-exp-computation.html).

Arguments:
- `j::Int`: index of the column of `C` we are interested in.
- `C::AbstractMatrix{T}`: cost matrix.
- `f::AbstractVector{T}`: dual potential.
- `log_α::AbstractVector{T}`: logarithm of the weights.
- `ε::T`: scaling parameter.

BrenierTwoFluid.value — Method

value

Compute the value of the Sinkhorn divergence associated with `S`, ``\sum_i α_i (f^\alpha_i - h^\alpha_i) + \sum_j β_j (f^\beta_j - h^\beta_j)``.

Arguments:
- `S::SinkhornDivergence`

Returns:
- `value(S)`

BrenierTwoFluid.y_gradient! — Method

y_gradient!

Compute the gradient of the Sinkhorn divergence with respect to the second probability distribution particle positions, divided by their weights, and store it in `S.V2.∇fh`.

Arguments:
- `S::SinkhornDivergence`

Returns:
- `∇S`: A reference to `S.V2.∇fh`.