BrenierTwoFluid Library Functions

BrenierTwoFluid.SinkhornBarycenter — Type

SinkhornBarycenter

A type representing a Sinkhorn barycenter problem.

BrenierTwoFluid.SinkhornDivergence — Type

SinkhornDivergence

Struct that holds two SinkhornVariables, a CostCollection, and SinkhornParameters. This is what most functions operate and dispatch on.

BrenierTwoFluid.SinkhornVariable — Type

SinkhornVariable

holds particle positions, weights, logarithm of the weights, potential and de-biasing potential, as well as those from the previous iteration

BrenierTwoFluid.compute! — Method

compute!(S::UnsafeSinkhornDivergence)

Solve the OT problem, i.e. do a bunch of Sinkhorn interations. Unsafe version.

BrenierTwoFluid.compute! — Method

compute!(S::SafeSinkhornDivergence)

Solve the OT problem, i.e. do a bunch of Sinkhorn interations. Safe version.

BrenierTwoFluid.sinkhorn_step! — Method

sinkhorn_step!

One iteration of the Sinkhorn algorithm.

BrenierTwoFluid.softmin — Method

softmin

Streaming log-sum-exp calculation